Revert "Merge remote-tracking branch 'origin/master'"
This reverts commit eb08f241b09b578a6bca6cd5bda9c177648880b8, reversing
changes made to a8902bf4714c5a0b134be00e70a4a2f623a212e1.

+++ VGP333_GravityJumperPhysicsTest/Assets/Scripts/FuelManager.cs
@@ -20,7 +20,7 @@ public class FuelManager : MonoBehaviour
 	{
 		if (psc != null)
 		{
-			FuelText.text = "Fuel = " + (int)psc.m_Fuel;
+			FuelText.text = "Fuel = " + psc.m_Fuel;
 		}
 	}
 }
